requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
google-generativeai==0.3.2
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
